Feat : DIG-101 루틴의 like, scrap 횟수 조회 로직 추가 및 변경 - 루틴 like, scrap 수 조회 로직 추가 및 변경 - DB에서 count하여 조회하는 방식으로 변경

+++ src/main/java/com/ogjg/daitgym/routine/dto/RoutineDto.java
@@ -11,12 +11,12 @@ public class RoutineDto {
     private String author;
     private String description;
     private boolean liked;
-    private int likeCounts;
-    private int scrapCounts;
+    private long likeCounts;
+    private long scrapCounts;
     private LocalDateTime createdAt;
 
     @Builder
-    public RoutineDto(Long id, String title, String author, String description, boolean liked, int likeCounts, int scrapCounts, LocalDateTime createdAt) {
+    public RoutineDto(Long id, String title, String author, String description, boolean liked, long likeCounts, long scrapCounts, LocalDateTime createdAt) {
         this.id = id;
         this.title = title;
         this.author = author;
